fix: add checksum validation when downloading the zookeeper client
